fixes issues
1. cannot find Drupal\path_alias\AliasRepository
