+++ tide_site.module
@@ -406,8 +406,8 @@ function tide_site_form_node_form_alter(&$form, FormStateInterface $form_state) 
   // Display all aliases of the current node.
   if (!$node->isNew()) {
     /** @var \Drupal\tide_site\AliasStorage $alias_storage */
-    $alias_storage = \Drupal::service('tide_site.alias_storage');
-    $aliases = $alias_storage->loadAll([
+    $alias_storage_helper = \Drupal::service('tide_site.alias_storage_helper');
+    $aliases = $alias_storage_helper->loadAll([
       'path' => '/node/' . $node->id(),
     ]);
     if ($aliases) {
@@ -542,12 +542,12 @@ function tide_site_taxonomy_term_delete($term) {
 function tide_site_node_update($node) {
   // Regenerate site aliases of the updated node.
   /** @var \Drupal\tide_site\AliasStorage $alias_storage */
-  $alias_storage = \Drupal::service('tide_site.alias_storage');
+  $alias_storage_helper = \Drupal::service('tide_site.alias_storage_helper');
 
   $path_storage = \Drupal::entityTypeManager()->getStorage('path_alias');
 
   /** @var \Drupal\node\NodeInterface $node */
-  if (!$alias_storage->loadAll(['path' => '/node/' . $node->id()])) {
+  if (!$alias_storage_helper->loadAll(['path' => '/node/' . $node->id()])) {
     /** @var \Drupal\tide_site\AliasStorageHelper $alias_helper */
     $alias_helper = \Drupal::service('tide_site.alias_storage_helper');
     $alias_helper->regenerateNodeSiteAliases($node);
